PLATUI-233: Moved Layout out into separate folder. Exposed all components through components import
