+++ src/main/play-26/twirl/uk/gov/hmrc/govukfrontend/views/layouts/govukLayout.scala.html
@@ -1,0 +1,95 @@
+@*
+ * Copyright 2019 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *@
+
+@import uk.gov.hmrc.govukfrontend.views.html.components._
+
+@this(
+        govukTemplate: GovukTemplate,
+        govukHeader: GovukHeader,
+        govukFooter: GovukFooter,
+        govukBackLink: GovukBackLink
+)
+
+@(pageTitle: Option[String] = None,
+        headBlock: Option[Html] = None,
+        headerBlock: Option[Html] = None,
+        beforeContentBlock: Option[Html] = None,
+        footerBlock: Option[Html] = None,
+        footerItems: Seq[FooterItem] = Seq.empty,
+        bodyEndBlock: Option[Html] = None,
+        scriptsBlock: Option[Html] = None
+)(contentBlock: Html)(implicit request: Request[_], messages: Messages)
+
+@headDefault = {
+@if(headBlock.isDefined) {
+    @headBlock
+}
+}
+
+@headerDefault = {
+@if(headerBlock.isDefined) {
+    @headerBlock
+} else {
+    @govukHeader(Header(
+        homepageUrl = Some(messages("service.homePageUrl")),
+        serviceName = Some(messages("service.name")),
+        serviceUrl = Some(messages("service.homePageUrl")),
+        containerClasses = "govuk-width-container"))
+}
+}
+
+@beforeContentDefault = {
+@if(beforeContentBlock.isDefined) {
+    @beforeContentBlock
+}
+}
+
+@footerDefault = {
+  @if(footerBlock.isDefined){
+    @footerBlock
+  }
+  @if(!footerItems.isEmpty) {
+      @govukFooter(new Footer(meta = Some(Meta(items = footerItems))))
+  }
+}
+
+@bodyEndDefault = {
+@if(bodyEndBlock.isDefined) {
+    @bodyEndBlock
+}
+@if(scriptsBlock.isDefined) {
+    @scriptsBlock
+}
+}
+
+@mainContentDefault = {
+    <div class="govuk-grid-row">
+        <div class="govuk-grid-column-two-thirds">
+        @contentBlock
+        </div>
+    </div>
+}
+
+@govukTemplate(
+    htmlLang = Some("en"),
+    pageTitle = pageTitle,
+    headBlock = Some(headDefault),
+    headerBlock = headerDefault,
+    beforeContentBlock = Some(beforeContentDefault),
+    footerBlock = footerDefault,
+    mainClasses = Some("govuk-main-wrapper--auto-spacing"),
+    bodyEndBlock = Some(bodyEndDefault))(mainContentDefault)
+
