+++ src/main/play-26/twirl/uk/gov/hmrc/govukfrontend/views/layouts/govukTemplate.scala.html
@@ -14,10 +14,10 @@
  * limitations under the License.
  *@
 
-    
 @import uk.gov.hmrc.govukfrontend.controllers.routes
+@import uk.gov.hmrc.govukfrontend.views.html.components._
 
-@this(govukHeader: govukHeader, govukFooter: govukFooter, govukSkipLink: govukSkipLink)
+@this(govukHeader: GovukHeader, govukFooter: GovukFooter, govukSkipLink: GovukSkipLink)
 
 @*FIXME use i18n support from Play instead of passing the language*@
 @(
